feat: move ParseError to thiserror
