+++ Cargo.toml
@@ -16,6 +16,7 @@ readme = "README.md"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 strum = { version = "0.28", features = ["derive"] }
+thiserror = "2.0"
 
 [build-dependencies]
 serde = { version = "1.0", features = ["derive"] }
