+++ src/v3.rs
@@ -5,8 +5,9 @@ use std::fmt;
 use std::str::FromStr;
 use strum::{Display, EnumString};
 
+use crate::error::ParseError;
 use crate::version::VersionV3;
-use crate::{ParseError, Severity as UnifiedSeverity};
+use crate::Severity as UnifiedSeverity;
 
 /// Represents a CVSS v3.0 or v3.1 score object.
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
