+++ src/v4_0/mod.rs
@@ -11,7 +11,8 @@ use std::fmt;
 use std::str::FromStr;
 use strum::{Display, EnumString};
 
-use crate::{ParseError, Severity as UnifiedSeverity};
+use crate::error::ParseError;
+use crate::Severity as UnifiedSeverity;
 
 /// Represents a CVSS v4.0 score object.
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
